change signature of DirectionalNodeSnapper, now returns a specially-encoded long and accepts only numeric parameters; encoded long can be decoded using static methods on DirectionalNodeSnapper, node offset is now a floating point rather than a double value (added precision of double is unnecessary), directionality of movement is now encoded in the value returned by the node snapper

+++ core/src/main/java/com/github/steanky/proxima/NodeHandler.java
@@ -5,5 +5,5 @@ import org.jetbrains.annotations.NotNull;
 
 @FunctionalInterface
 public interface NodeHandler {
-    void handle(@NotNull Node node, int x, int y, int z, double yOffset);
+    void handle(@NotNull Node node, int x, int y, int z, float yOffset);
 }
